User Association - Create Tasks for Logged-in Users

+++ task-management/src/tasks/controller.py
@@ -13,8 +13,8 @@ def create_task(body:TaskSchema,db : Session,user:UserModel ):
     db.refresh(new_tasks)
     return new_tasks
 
-def get_tasks(db:Session):
-    tasks = db.query(TaskModel).all()
+def get_tasks(db:Session,user:UserModel):
+    tasks = db.query(TaskModel).filter(TaskModel.user_id == user.id).all()
     return tasks
 
 
@@ -26,10 +26,13 @@ def get_one_task(task_id:int,db:Session):
 
     return one_task
 
-def update_task(body:TaskSchema,task_id:int,db:Session):
+def update_task(body:TaskSchema,task_id:int,db:Session,user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404, detail="Task id not found in database")
+    
+    if one_task.user_id != user.id:
+        raise HTTPException(404, detail="YOu are not allowed to update this task.")
     
     body = body.model_dump()
 
@@ -46,10 +49,13 @@ def update_task(body:TaskSchema,task_id:int,db:Session):
     return one_task
 
 
-def delete_task(task_id:int,db:Session):
+def delete_task(task_id:int,db:Session,user:UserModel):
     one_task = db.query(TaskModel).get(task_id)
     if not one_task:
         raise HTTPException(404,detail="Task id not found!")
+    
+    if one_task.user_id != user.id:
+        raise HTTPException(404, detail="YOu are not allowed to delete this task.")
 
     db.delete(one_task)
     db.commit()
